darwinLintPackage

Darwin lintr object, using default lintr object with camelCase

Usage

darwinLintPackage()

Value

List of lint objects.

Examples

darwinLintPackage()
#> R\checkDependencies.R:30:32: style: [infix_spaces_linter] Put spaces around all infix operators.
#>     dplyr::rename("version_rec"="version") %>%
#>                                ^
#> R\checkDependencies.R:49:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   not_permitted <- dependencies %>%
#>   ^~~~~~~~~~~~~
#> R\checkDependencies.R:49:3: warning: [object_usage_linter] local variable 'not_permitted' assigned but may not be used
#>   not_permitted <- dependencies %>%
#>   ^~~~~~~~~~~~~
#> R\checkDependencies.R:67:34: style: [object_name_linter] Variable and function name style should be camelCase.
#> messagePermission <- function(i, not_permitted) {
#>                                  ^~~~~~~~~~~~~
#> R\checkDependencies.R:113:5: style: [spaces_left_parentheses_linter] Place a space before left parenthesis, except in a function call.
#>   if(is.null(packageName)) {
#>     ^
#> R\checkDependencies.R:126:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   not_permitted <- getNotPermitted(dependencies, permittedPackages)
#>   ^~~~~~~~~~~~~
#> R\checkDependencies.R:128:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   n_not_permitted <- length(not_permitted)
#>   ^~~~~~~~~~~~~~~
#> R\checkDependencies.R:134:5: style: [spaces_left_parentheses_linter] Place a space before left parenthesis, except in a function call.
#>   if(n_not_permitted == 0) {
#>     ^
#> R\checkDependencies.R:157:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>             dplyr::select(.data$package, version, date, .data$downloads_last_month,
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~
#> R\checkDependencies.R:161:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>     example$url <- stringr::str_replace_all(string = example$url, pattern = ",\\\n", replacement = " ")
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:164:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>     "{.emph Please create a new issue at https://github.com/mvankessel-EMC/DependencyReviewerWhitelists/
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:178:4: style: [commented_code_linter] Commented code should be removed.
#>   #n_diffVersions <- length(diffVersions$package)
#>    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:189:7: style: [commented_code_linter] Commented code should be removed.
#>   #   cli::cli_div(theme = list (.alert = list(color = "red")))
#>       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:193:7: style: [commented_code_linter] Commented code should be removed.
#>   #   cli::cli_end()
#>       ^~~~~~~~~~~~~~
#> R\checkDependencies.R:200:7: style: [commented_code_linter] Commented code should be removed.
#>   #   cli::cli_alert_warning("Please require recommended versions")
#>       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\getDefaultPermittedPackages.R:52:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       file = "https://raw.githubusercontent.com/mvankessel-EMC/DependencyReviewerWhitelists/main/dependencies.csv",
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\getDefaultPermittedPackages.R:79:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       file = "https://raw.githubusercontent.com/OHDSI/Hades/main/extras/packages.csv",
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~
#> R\getDefaultPermittedPackages.R:81:45: style: [pipe_continuation_linter] `%>%` should always have a space before it and a new line after it, unless the full pipeline fits on one line.
#>       header = TRUE) %>% select(.data$name) %>%
#>                                             ^~~
#> R\getGraphData.R:8:81: style: [line_length_linter] Lines should not be more than 80 characters.
#> #'     Types of packages to be included in the result. Default: c("imports", "depends")
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~
#> R\getGraphData.R:9:81: style: [line_length_linter] Lines should not be more than 80 characters.
#> #'     Available types are: "imports", "depends", "suggests", "enhances", "linkingto"
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~
#> R\getGraphData.R:22:47: style: [object_name_linter] Variable and function name style should be camelCase.
#> getGraphData <- function(path = here::here(), excluded_packages = c(""), package_types = c("imports", "depends")) {
#>                                               ^~~~~~~~~~~~~~~~~
#> R\getGraphData.R:22:74: style: [object_name_linter] Variable and function name style should be camelCase.
#> getGraphData <- function(path = here::here(), excluded_packages = c(""), package_types = c("imports", "depends")) {
#>                                                                          ^~~~~~~~~~~~~
#> R\getGraphData.R:22:81: style: [line_length_linter] Lines should not be more than 80 characters.
#> getGraphData <- function(path = here::here(), excluded_packages = c(""), package_types = c("imports", "depends")) {
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\getGraphData.R:30:9: warning: [seq_linter] 1:nrow(...) is likely to be wrong in the empty edge case. Use seq_len() instead.
#>     X = 1:nrow(fData),
#>         ^~~~~~~~~~~~~
#> R\getGraphData.R:37:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   pkg_deps <- dplyr::bind_rows(lapply(X = 1:nrow(fData), FUN = function(row) {
#>   ^~~~~~~~
#> R\getGraphData.R:37:43: warning: [seq_linter] 1:nrow(...) is likely to be wrong in the empty edge case. Use seq_len() instead.
#>   pkg_deps <- dplyr::bind_rows(lapply(X = 1:nrow(fData), FUN = function(row) {
#>                                           ^~~~~~~~~~~~~
#> R\getGraphData.R:44:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   pkg_deps <- pkg_deps %>%
#>   ^~~~~~~~
#> R\getGraphData.R:48:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   net_data <- tidygraph::as_tbl_graph(
#>   ^~~~~~~~
#> R\getGraphData.R:48:3: warning: [object_usage_linter] local variable 'net_data' assigned but may not be used
#>   net_data <- tidygraph::as_tbl_graph(
#>   ^~~~~~~~
#> R\getGraphData.R:52:1: style: [trailing_blank_lines_linter] Trailing blank lines are superfluous.
#> 
#> ^
#> R\summariseFunctionUse.R:33:28: style: [object_name_linter] Variable and function name style should be camelCase.
#> funsUsedInLine <- function(file_txt, file_name, i, verbose = FALSE) {
#>                            ^~~~~~~~
#> R\summariseFunctionUse.R:33:38: style: [object_name_linter] Variable and function name style should be camelCase.
#> funsUsedInLine <- function(file_txt, file_name, i, verbose = FALSE) {
#>                                      ^~~~~~~~~
#> R\summariseFunctionUse.R:42:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     fun_vec <- unlist(stringr::str_extract_all(
#>     ^~~~~~~
#> R\summariseFunctionUse.R:46:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     fun_vec <- stringr::str_remove_all(
#>     ^~~~~~~
#> R\summariseFunctionUse.R:50:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     fun_vec <- stringr::str_split(
#>     ^~~~~~~
#> R\summariseFunctionUse.R:54:7: style: [spaces_left_parentheses_linter] Place a space before left parenthesis, except in a function call.
#>     if(length(fun_vec) > 0) {
#>       ^
#> R\summariseFunctionUse.R:55:7: style: [object_name_linter] Variable and function name style should be camelCase.
#>       fun_vec <- lapply(
#>       ^~~~~~~
#> R\summariseFunctionUse.R:58:13: style: [spaces_left_parentheses_linter] Place a space before left parenthesis, except in a function call.
#>           if(length(x) == 1) {
#>             ^
#> R\summariseFunctionUse.R:73:9: style: [spaces_left_parentheses_linter] Place a space before left parenthesis, except in a function call.
#>       if(verbose == TRUE) {
#>         ^
#> R\summariseFunctionUse.R:93:52: style: [object_name_linter] Variable and function name style should be camelCase.
#> funsUsedInFile <- function(files, verbose = FALSE, in_package = TRUE) {
#>                                                    ^~~~~~~~~~
#> R\summariseFunctionUse.R:95:7: style: [spaces_left_parentheses_linter] Place a space before left parenthesis, except in a function call.
#>     if(verbose) {
#>       ^
#> R\summariseFunctionUse.R:99:7: style: [spaces_left_parentheses_linter] Place a space before left parenthesis, except in a function call.
#>     if(in_package) {
#>       ^
#> R\summariseFunctionUse.R:100:7: style: [object_name_linter] Variable and function name style should be camelCase.
#>       file_txt <- readLines(here::here("R", file))
#>       ^~~~~~~~
#> R\summariseFunctionUse.R:102:7: style: [object_name_linter] Variable and function name style should be camelCase.
#>       file_txt <- readLines(file)
#>       ^~~~~~~~
#> R\summariseFunctionUse.R:105:5: warning: [object_usage_linter] local variable 'out' assigned but may not be used
#>     out <- sapply(
#>     ^~~
#> R\summariseFunctionUse.R:106:11: warning: [seq_linter] 1:length(...) is likely to be wrong in the empty edge case. Use seq_along() instead.
#>       X = 1:length(file_txt),
#>           ^~~~~~~~~~~~~~~~~~
#> R\summariseFunctionUse.R:141:12: style: [object_name_linter] Variable and function name style should be camelCase.
#>   function(r_files = list.files(here::here("R")),
#>            ^~~~~~~
#> R\summariseFunctionUse.R:143:12: style: [object_name_linter] Variable and function name style should be camelCase.
#>            in_package = TRUE) {
#>            ^~~~~~~~~~
#> R\summariseFunctionUse.R:145:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     deps_used <- funsUsedInFile(r_files, verbose, in_package)
#>     ^~~~~~~~~
#> R\summariseFunctionUse.R:147:9: style: [commented_code_linter] Commented code should be removed.
#>     #   stop(paste(r_files, "not found"))
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\summariseFunctionUse.R:152:7: style: [object_name_linter] Variable and function name style should be camelCase.
#>       deps_used <- tibble(
#>       ^~~~~~~~~
#> R\summariseFunctionUse.R:160:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     deps_used <- dplyr::bind_rows(deps_used) %>%
#>     ^~~~~~~~~
#> R\summariseFunctionUse.R:164:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     deps_used$pkg[deps_used$fun %in% ls("package:base")] <- "base"
#>     ^~~~~~~~~
#> R\summariseFunctionUse.R:167:1: style: [trailing_blank_lines_linter] Trailing blank lines are superfluous.
#> 
#> ^
#> inst\shinyApp\server.R:22:5: style: [commented_code_linter] Commented code should be removed.
#>   # DependencyReviewer::darwinLintScore(DependencyReviewer::darwinLintPackage)
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:40:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       DependencyReviewer::summariseFunctionUse(r_files = list.files(here::here("R"))) %>%
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~
#> inst\shinyApp\server.R:51:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       DependencyReviewer::summariseFunctionUse(r_files = list.files(here::here("R")))
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~
#> inst\shinyApp\server.R:71:31: style: [object_name_linter] Variable and function name style should be camelCase.
#>                               data_sub <- data %>%
#>                               ^~~~~~~~
#> inst\shinyApp\server.R:78:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>                                               mapping = ggplot2::aes(x = fun, y = n, fill = pkg)) +
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:85:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>                                                axis.text.x = (ggplot2::element_text(
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~
#> inst\shinyApp\server.R:157:48: style: [commas_linter] Commas should always have a space after.
#>     options <- names(tail(igraph::V(graphData),-1))
#>                                                ^
#> inst\shinyApp\server.R:203:7: style: [commented_code_linter] Commented code should be removed.
#>   #   out <- DependencyReviewer::darwinLintScore(DependencyReviewer::darwinLintPackage)
#>       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:203:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>   #   out <- DependencyReviewer::darwinLintScore(DependencyReviewer::darwinLintPackage)
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~
#> inst\shinyApp\server.R:204:7: style: [commented_code_linter] Commented code should be removed.
#>   #   return(out)
#>       ^~~~~~~~~~~
#> inst\shinyApp\server.R:209:9: style: [commented_code_linter] Commented code should be removed.
#>   #     shiny::setProgress(value = 0, message = "Running Lintr on package")
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:211:9: style: [commented_code_linter] Commented code should be removed.
#>   #     out <- DT::datatable(getLintrTable())
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:213:9: style: [commented_code_linter] Commented code should be removed.
#>   #     shiny::setProgress(value = 1, message = "Supplying output in table")
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:215:7: style: [commented_code_linter] Commented code should be removed.
#>   #   return(out)
#>       ^~~~~~~~~~~
#> inst\shinyApp\ui.R:31:13: style: [commented_code_linter] Commented code should be removed.
#>       #     choices = c("ALL", list.files(here::here("R")))
#>             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:38:13: style: [commented_code_linter] Commented code should be removed.
#>       #     choices = c("base", "unknown")
#>             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:49:17: style: [commented_code_linter] Commented code should be removed.
#>           #     DT::dataTableOutput("lintrTable")
#>                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:104:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>                   choices = c("imports", "depends", "suggests", "enhances", "linkingto")
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~
#> inst\testScript.R:45:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       file = "https://raw.githubusercontent.com/mvankessel-EMC/DependencyReviewerWhitelists/main/dependencies.csv",
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\testScript.R:52:21: warning: [object_usage_linter] no visible binding for global variable 'Package'
#>       dplyr::select(Package, Built) %>%
#>                     ^~~~~~~
#> inst\testScript.R:52:21: warning: [object_usage_linter] no visible binding for global variable 'Package'
#>       dplyr::select(Package, Built) %>%
#>                     ^~~~~~~
#> inst\testScript.R:52:30: warning: [object_usage_linter] no visible binding for global variable 'Built'
#>       dplyr::select(Package, Built) %>%
#>                              ^~~~~
#> inst\testScript.R:52:30: warning: [object_usage_linter] no visible binding for global variable 'Built'
#>       dplyr::select(Package, Built) %>%
#>                              ^~~~~
#> inst\testScript.R:70:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       file = "https://raw.githubusercontent.com/OHDSI/Hades/main/extras/packages.csv",
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~
#> inst\testScript.R:72:33: warning: [object_usage_linter] no visible binding for global variable 'name'
#>       header = TRUE) %>% select(name) %>%
#>                                 ^~~~
#> inst\testScript.R:72:33: warning: [object_usage_linter] no visible binding for global variable 'name'
#>       header = TRUE) %>% select(name) %>%
#>                                 ^~~~
#> inst\testScript.R:72:39: style: [pipe_continuation_linter] `%>%` should always have a space before it and a new line after it, unless the full pipeline fits on one line.
#>       header = TRUE) %>% select(name) %>%
#>                                       ^~~
#> inst\testScript.R:90:16: warning: [object_usage_linter] no visible binding for global variable 'package'
#>         select(package, version))
#>                ^~~~~~~
#> tests\testthat\test-checkDependencies.R:6:5: style: [commented_code_linter] Commented code should be removed.
#> #   expect_message(checkDependencies())
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> tests\testthat\test-darwinLint.R:6:5: style: [commented_code_linter] Commented code should be removed.
#> #   expect_type(darwinLintScore(lintFunction = darwinLintPackage), "list")
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> tests\testthat\test-getDefaultPermittedPackages.R:6:5: style: [commented_code_linter] Commented code should be removed.
#> #   expect_type(getDefaultPermittedPackages(), "list")
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> vignettes\Documentation.Rmd:219:20: style: [trailing_whitespace_linter] Trailing whitespace is superfluous.
#>   data = funCounts, 
#>                    ^
#> vignettes\Documentation.Rmd:223:15: style: [trailing_whitespace_linter] Trailing whitespace is superfluous.
#>     vars(pkg), 
#>               ^
#> vignettes\Documentation.Rmd:224:23: style: [trailing_whitespace_linter] Trailing whitespace is superfluous.
#>     scales = "free_x", 
#>                       ^