darwinLintPackage

Darwin lintr object, using default lintr object with camelCase

darwinLintPackage()

Value

List of lint objects.

Examples

darwinLintPackage()
#> R\checkDependencies.R:50:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   not_permitted <- dependencies %>%
#>   ^~~~~~~~~~~~~
#> R\checkDependencies.R:50:3: warning: [object_usage_linter] local variable 'not_permitted' assigned but may not be used
#>   not_permitted <- dependencies %>%
#>   ^~~~~~~~~~~~~
#> R\checkDependencies.R:69:34: style: [object_name_linter] Variable and function name style should be camelCase.
#> messagePermission <- function(i, not_permitted) {
#>                                  ^~~~~~~~~~~~~
#> R\checkDependencies.R:126:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   not_permitted <- getNotPermitted(dependencies, permittedPackages)
#>   ^~~~~~~~~~~~~
#> R\checkDependencies.R:128:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   n_not_permitted <- length(not_permitted)
#>   ^~~~~~~~~~~~~~~
#> R\checkDependencies.R:169:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>     example$url <- stringr::str_replace_all(string = example$url, pattern = ",\\\n", replacement = " ")
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:172:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       "{.emph Please create a new issue at https://github.com/mvankessel-EMC/DependencyReviewerWhitelists/
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:183:6: style: [commented_code_linter] Commented code should be removed.
#>   #  dependencies = dependencies,
#>      ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:186:5: style: [commented_code_linter] Commented code should be removed.
#>   # n_diffVersions <- length(diffVersions$package)
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:197:7: style: [commented_code_linter] Commented code should be removed.
#>   #   cli::cli_div(theme = list (.alert = list(color = "red")))
#>       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:201:7: style: [commented_code_linter] Commented code should be removed.
#>   #   cli::cli_end()
#>       ^~~~~~~~~~~~~~
#> R\checkDependencies.R:204:9: style: [commented_code_linter] Commented code should be removed.
#>   #     X = 1:n_diffVersions,
#>         ^~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:205:9: style: [commented_code_linter] Commented code should be removed.
#>   #     FUN = messagePackageVersion,
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\checkDependencies.R:208:7: style: [commented_code_linter] Commented code should be removed.
#>   #   cli::cli_alert_warning("Please require recommended versions")
#>       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\getDefaultPermittedPackages.R:54:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       file = "https://raw.githubusercontent.com/mvankessel-EMC/DependencyReviewerWhitelists/main/dependencies.csv",
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\getDefaultPermittedPackages.R:84:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       file = "https://raw.githubusercontent.com/OHDSI/Hades/main/extras/packages.csv",
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~
#> R\getGraphData.R:8:81: style: [line_length_linter] Lines should not be more than 80 characters.
#> #'     Types of packages to be included in the result. Default: c("imports", "depends")
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~
#> R\getGraphData.R:9:81: style: [line_length_linter] Lines should not be more than 80 characters.
#> #'     Available types are: "imports", "depends", "suggests", "enhances", "linkingto"
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~
#> R\getGraphData.R:22:47: style: [object_name_linter] Variable and function name style should be camelCase.
#> getGraphData <- function(path = here::here(), excluded_packages = c(""), package_types = c("imports", "depends")) {
#>                                               ^~~~~~~~~~~~~~~~~
#> R\getGraphData.R:22:74: style: [object_name_linter] Variable and function name style should be camelCase.
#> getGraphData <- function(path = here::here(), excluded_packages = c(""), package_types = c("imports", "depends")) {
#>                                                                          ^~~~~~~~~~~~~
#> R\getGraphData.R:22:81: style: [line_length_linter] Lines should not be more than 80 characters.
#> getGraphData <- function(path = here::here(), excluded_packages = c(""), package_types = c("imports", "depends")) {
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\getGraphData.R:30:9: warning: [seq_linter] 1:nrow(...) is likely to be wrong in the empty edge case. Use seq_len(nrow(...)) instead.
#>     X = 1:nrow(fData),
#>         ^~~~~~~~~~~~~
#> R\getGraphData.R:38:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   pkg_deps <- dplyr::bind_rows(lapply(X = 1:nrow(fData), FUN = function(row) {
#>   ^~~~~~~~
#> R\getGraphData.R:38:43: warning: [seq_linter] 1:nrow(...) is likely to be wrong in the empty edge case. Use seq_len(nrow(...)) instead.
#>   pkg_deps <- dplyr::bind_rows(lapply(X = 1:nrow(fData), FUN = function(row) {
#>                                           ^~~~~~~~~~~~~
#> R\getGraphData.R:45:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   pkg_deps <- pkg_deps %>%
#>   ^~~~~~~~
#> R\getGraphData.R:49:3: style: [object_name_linter] Variable and function name style should be camelCase.
#>   net_data <- tidygraph::as_tbl_graph(
#>   ^~~~~~~~
#> R\getGraphData.R:49:3: warning: [object_usage_linter] local variable 'net_data' assigned but may not be used
#>   net_data <- tidygraph::as_tbl_graph(
#>   ^~~~~~~~
#> R\summariseFunctionUse.R:33:28: style: [object_name_linter] Variable and function name style should be camelCase.
#> funsUsedInLine <- function(file_txt, file_name, i, verbose = FALSE) {
#>                            ^~~~~~~~
#> R\summariseFunctionUse.R:33:38: style: [object_name_linter] Variable and function name style should be camelCase.
#> funsUsedInLine <- function(file_txt, file_name, i, verbose = FALSE) {
#>                                      ^~~~~~~~~
#> R\summariseFunctionUse.R:45:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     fun_vec <- unlist(stringr::str_extract_all(
#>     ^~~~~~~
#> R\summariseFunctionUse.R:50:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     fun_vec <- stringr::str_remove_all(
#>     ^~~~~~~
#> R\summariseFunctionUse.R:55:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     fun_vec <- stringr::str_split(
#>     ^~~~~~~
#> R\summariseFunctionUse.R:61:7: style: [object_name_linter] Variable and function name style should be camelCase.
#>       fun_vec <- lapply(
#>       ^~~~~~~
#> R\summariseFunctionUse.R:99:52: style: [object_name_linter] Variable and function name style should be camelCase.
#> funsUsedInFile <- function(files, verbose = FALSE, in_package = TRUE) {
#>                                                    ^~~~~~~~~~
#> R\summariseFunctionUse.R:106:7: style: [object_name_linter] Variable and function name style should be camelCase.
#>       file_txt <- readLines(here::here("R", file))
#>       ^~~~~~~~
#> R\summariseFunctionUse.R:108:7: style: [object_name_linter] Variable and function name style should be camelCase.
#>       file_txt <- readLines(file)
#>       ^~~~~~~~
#> R\summariseFunctionUse.R:111:5: warning: [object_usage_linter] local variable 'out' assigned but may not be used
#>     out <- sapply(
#>     ^~~
#> R\summariseFunctionUse.R:112:11: warning: [seq_linter] 1:length(...) is likely to be wrong in the empty edge case. Use seq_along(...) instead.
#>       X = 1:length(file_txt),
#>           ^~~~~~~~~~~~~~~~~~
#> R\summariseFunctionUse.R:149:12: style: [object_name_linter] Variable and function name style should be camelCase.
#>   function(r_files = list.files(here::here("R")),
#>            ^~~~~~~
#> R\summariseFunctionUse.R:151:12: style: [object_name_linter] Variable and function name style should be camelCase.
#>            in_package = TRUE) {
#>            ^~~~~~~~~~
#> R\summariseFunctionUse.R:153:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     deps_used <- funsUsedInFile(r_files, verbose, in_package)
#>     ^~~~~~~~~
#> R\summariseFunctionUse.R:155:9: style: [commented_code_linter] Commented code should be removed.
#>     #   stop(paste(r_files, "not found"))
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> R\summariseFunctionUse.R:160:7: style: [object_name_linter] Variable and function name style should be camelCase.
#>       deps_used <- tibble(
#>       ^~~~~~~~~
#> R\summariseFunctionUse.R:168:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     deps_used <- dplyr::bind_rows(deps_used) %>%
#>     ^~~~~~~~~
#> R\summariseFunctionUse.R:172:5: style: [object_name_linter] Variable and function name style should be camelCase.
#>     deps_used$pkg[deps_used$fun %in% ls("package:base")] <- "base"
#>     ^~~~~~~~~
#> inst\shinyApp\server.R:22:5: style: [commented_code_linter] Commented code should be removed.
#>   # DependencyReviewer::darwinLintScore(DependencyReviewer::darwinLintPackage)
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:40:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       DependencyReviewer::summariseFunctionUse(r_files = list.files(here::here("R"))) %>%
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~
#> inst\shinyApp\server.R:51:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       DependencyReviewer::summariseFunctionUse(r_files = list.files(here::here("R")))
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~
#> inst\shinyApp\server.R:71:31: style: [object_name_linter] Variable and function name style should be camelCase.
#>                               data_sub <- data %>%
#>                               ^~~~~~~~
#> inst\shinyApp\server.R:78:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>                                               mapping = ggplot2::aes(x = fun, y = n, fill = pkg)) +
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:85:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>                                                axis.text.x = (ggplot2::element_text(
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~
#> inst\shinyApp\server.R:157:48: style: [commas_linter] Commas should always have a space after.
#>     options <- names(tail(igraph::V(graphData),-1))
#>                                                ^
#> inst\shinyApp\server.R:203:7: style: [commented_code_linter] Commented code should be removed.
#>   #   out <- DependencyReviewer::darwinLintScore(DependencyReviewer::darwinLintPackage)
#>       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:203:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>   #   out <- DependencyReviewer::darwinLintScore(DependencyReviewer::darwinLintPackage)
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~
#> inst\shinyApp\server.R:204:7: style: [commented_code_linter] Commented code should be removed.
#>   #   return(out)
#>       ^~~~~~~~~~~
#> inst\shinyApp\server.R:209:9: style: [commented_code_linter] Commented code should be removed.
#>   #     shiny::setProgress(value = 0, message = "Running Lintr on package")
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:211:9: style: [commented_code_linter] Commented code should be removed.
#>   #     out <- DT::datatable(getLintrTable())
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:213:9: style: [commented_code_linter] Commented code should be removed.
#>   #     shiny::setProgress(value = 1, message = "Supplying output in table")
#>         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\server.R:215:7: style: [commented_code_linter] Commented code should be removed.
#>   #   return(out)
#>       ^~~~~~~~~~~
#> inst\shinyApp\ui.R:27:11: style: [commented_code_linter] Commented code should be removed.
#>       #   width = 3,
#>           ^~~~~~~~~~
#> inst\shinyApp\ui.R:29:13: style: [commented_code_linter] Commented code should be removed.
#>       #     inputId = "file",
#>             ^~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:30:13: style: [commented_code_linter] Commented code should be removed.
#>       #     label = "File",
#>             ^~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:31:13: style: [commented_code_linter] Commented code should be removed.
#>       #     choices = c("ALL", list.files(here::here("R")))
#>             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:35:13: style: [commented_code_linter] Commented code should be removed.
#>       #     inline = TRUE,
#>             ^~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:36:13: style: [commented_code_linter] Commented code should be removed.
#>       #     inputId = "excludes",
#>             ^~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:37:13: style: [commented_code_linter] Commented code should be removed.
#>       #     label = "Exclude Packages",
#>             ^~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:38:13: style: [commented_code_linter] Commented code should be removed.
#>       #     choices = c("base", "unknown")
#>             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:49:17: style: [commented_code_linter] Commented code should be removed.
#>           #     DT::dataTableOutput("lintrTable")
#>                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\shinyApp\ui.R:104:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>                   choices = c("imports", "depends", "suggests", "enhances", "linkingto")
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~
#> inst\testScript.R:45:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       file = "https://raw.githubusercontent.com/mvankessel-EMC/DependencyReviewerWhitelists/main/dependencies.csv",
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> inst\testScript.R:52:21: warning: [object_usage_linter] no visible binding for global variable 'Package'
#>       dplyr::select(Package, Built) %>%
#>                     ^~~~~~~
#> inst\testScript.R:52:21: warning: [object_usage_linter] no visible binding for global variable 'Package'
#>       dplyr::select(Package, Built) %>%
#>                     ^~~~~~~
#> inst\testScript.R:52:30: warning: [object_usage_linter] no visible binding for global variable 'Built'
#>       dplyr::select(Package, Built) %>%
#>                              ^~~~~
#> inst\testScript.R:52:30: warning: [object_usage_linter] no visible binding for global variable 'Built'
#>       dplyr::select(Package, Built) %>%
#>                              ^~~~~
#> inst\testScript.R:70:81: style: [line_length_linter] Lines should not be more than 80 characters.
#>       file = "https://raw.githubusercontent.com/OHDSI/Hades/main/extras/packages.csv",
#> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~
#> inst\testScript.R:72:33: warning: [object_usage_linter] no visible binding for global variable 'name'
#>       header = TRUE) %>% select(name) %>%
#>                                 ^~~~
#> inst\testScript.R:72:33: warning: [object_usage_linter] no visible binding for global variable 'name'
#>       header = TRUE) %>% select(name) %>%
#>                                 ^~~~
#> inst\testScript.R:72:39: style: [pipe_continuation_linter] `%>%` should always have a space before it and a new line after it, unless the full pipeline fits on one line.
#>       header = TRUE) %>% select(name) %>%
#>                                       ^~~
#> inst\testScript.R:90:16: warning: [object_usage_linter] no visible binding for global variable 'package'
#>         select(package, version))
#>                ^~~~~~~
#> tests\testthat\test-checkDependencies.R:6:5: style: [commented_code_linter] Commented code should be removed.
#> #   expect_message(checkDependencies())
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> tests\testthat\test-darwinLint.R:6:5: style: [commented_code_linter] Commented code should be removed.
#> #   expect_type(darwinLintScore(lintFunction = darwinLintPackage), "list")
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> tests\testthat\test-getDefaultPermittedPackages.R:6:5: style: [commented_code_linter] Commented code should be removed.
#> #   expect_type(getDefaultPermittedPackages(), "list")
#>     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~